added structure to app.js made a drawingboard page.
