+++ client/src/App.js
@@ -1,54 +1,24 @@
-import React, { useState } from "react";
+import React from "react";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import "./App.css";
-// Can switch palette to react-color
-// import { GithubPicker } from "react-color";
-import Palette from "./components/Palette";
 
-// Import Templates
-import StarTrio from "./components/StarTrio.js";
+
+import DrawingBoard from "./pages/DrawingBoard";
 
 function App() {
-  //For react-color front test
-  // const [currentColour, setColour] = useState("#B80000");
-
-  // Path's get color filled
-  const [fillColours, setFillColours] = useState(Array(10).fill("white"));
-  
-  // Current colour, How to set new colour
-  const [currentColour, setCurrentColour] = useState("blue");
-
-  // How to change Colour
-  const onFillColour = (i) => {
-    let newFillColours = fillColours.slice(0);
-    newFillColours[i] = currentColour;
-    setFillColours(newFillColours);
-
-  };
   return (
     <div className="App">
-      <header>
-        <h1>
-          {/* style color can not have a u */}
-          <p style={{ color: currentColour }}>Colour</p>
-          <p style={{ color: currentColour }}>Harmony</p>
-          <p style={{ color: currentColour }}>Blend</p>
-        </h1>
-      </header>
-      {/* This will eventually be a switch load case */}
-        <StarTrio fillColours={fillColours} onFill={onFillColour} />
-        <Palette currentColour={currentColour} changeColour={setCurrentColour}  />
-        {/* We may want to switch to react-color
-        <GithubPicker
-          color={currentColour}
-          onChangeComplete={(colour) => {
-            setColour(colour.hex);
-          }}
-          width="200px"
-          colours={['#B80000', '#DB3E00', '#FCCB00', '#008B02', '#006B76', '#1273DE', '#004DCF', '#5300EB', '#EB9694', '#FAD0C3', '#FEF3BD', '#C1E1C5', '#BEDADC', '#C4DEF6', '#BED3F3', '#D4C4FB']}
-          triangle="hide"
-          // onSwatchHover />
-         */}
-      </div>
+      {/* <ApolloProvider> */}
+      <Router>
+        <Switch>
+          {/* <Route component={Login} />
+              <Route component={Signup}/> 
+          */}
+          <Route exact path="/drawingBoard" component={DrawingBoard} />
+        </Switch>
+      </Router>
+      {/* </ApolloProvider> */}
+    </div>
   );
 }
