+++ client/src/components/Palette/index.js
