+++ client/src/components/StarTrio/index.js
